refactored sif import tests, reduce repeating code in upload.js, increased coverage of stray data, blank/one node sif files handled
